third_party/blink/web_tests/platform/generic/external/wpt/encoding-detection/vi-windows-1258.tentative-expected.txt

This is a testharness.js-based test.
FAIL Check detection result assert_equals: Expected windows-1258 expected "windows-1258" but got "windows-1252"
Harness: the test ran to completion.
